Store component information in a data structure
This will allow the structure for the generated program (i.e. which
files belong to which components, etc.) to be deduced once and then used
where ever needed (generating files and generating build configuration
file).

Note that there were a couple of workarounds because of what seemed like
compiler bugs. They may be just a reflection of my Fortran ignorance as
well.