Add performence tests and fix problems they unrevealed
Specifically fix the problem where a statically allocated string buffer
was used to store the string for build edges. This buffer ran out when
sufficient number of files were added into a component (thus creating
equal number of dependencies from the main file of the component to the
other files).