Use less stack space for storing program structure
The file names are now generated on the fly, rather than storing each
and every one of them in stack memory.

This is not actually enough to make the generation able to generate
large number of files (already 200 fails). This is because
BuildConfiguration is using also very large amounts of stack memory.

That needs to be fixed separately.