Support larger component sizes
Specifically this avoids stack overflow when generating ninja file.

I am not fully happy about the change. The original implementation had
functions returning arrays of strings and once the whole contents was
available it was written into the file.

The original code was fairly clean and did not have to deal with error
situations. But it caused stack overflows, because (?) the strings were
stored in allocatable arrays, which are stored in stack (?).

The new implementation writes rows immediately to file. It works, but
for every file operation you need to check for the error.

Note: I did not add automated performance test that proved the problem.
The reason is that the test got stuck with the test. My hypothesis is
that there is so much output from ninja that the pipe handling in Ruby
gets choked (I have no real evidence on this; I've just seen similar
things in Python side, and I when I killed a stuck ninja process, the
output was cropped after a while even though all the processes had
successfully finished). Thus: tested only manually with 1000 files.